requirements.txt
dateutil.relativedelta
sqlalchemy
psycopg2-binary
bs4
pandas
numpy
requests
argparse